feat: add Course, CourseEdition, and Enrollment models with initial implementation
