+++ backend/apps/courses/models/__init__.py
@@ -1,0 +1,6 @@
+"""Course app models package."""
+
+from .course import Course, CourseEdition
+from .enrollment import Enrollment
+
+__all__ = ["Course", "CourseEdition", "Enrollment"]
